Linux: several keyboard input improvements.
- No longer assumes a US-QWERTY layout (uses SDL equivalent of Windows VK_ values, not raw keyboard scancodes)
- Fixes various keys not working for odd reasons (we lost bits downcasting from uint32 to uint16, causing key values to change or overlap other keys).

(Fixes UE-11582 and UE-14143 and probably others.)

Misc:
- Reverted to using WarpMouseInWindow due to reports that Global is not always reliable.
- Added a workaround for relative mouse movement being wrong in some cases.

#platformnotify Josh.Adams
#codereview Ryan.Gordon, Josh.Adams, Marc.Audy

[CL 2625112 by Dmitry Rekman in Main branch]

+++ Engine/Source/Runtime/Core/Private/HTML5/HTML5PlatformMisc.cpp
@@ -38,12 +38,12 @@ GenericApplication* FHTML5Misc::CreateApplication()
 	return FHTML5Application::CreateHTML5Application();
 }
 
-uint32 FHTML5Misc::GetCharKeyMap(uint16* KeyCodes, FString* KeyNames, uint32 MaxMappings)
+uint32 FHTML5Misc::GetCharKeyMap(uint32* KeyCodes, FString* KeyNames, uint32 MaxMappings)
 {
 	return FGenericPlatformMisc::GetStandardPrintableKeyMap(KeyCodes, KeyNames, MaxMappings, false, true);
 }
 
-uint32 FHTML5Misc::GetKeyMap(uint16* KeyCodes, FString* KeyNames, uint32 MaxMappings)
+uint32 FHTML5Misc::GetKeyMap(uint32* KeyCodes, FString* KeyNames, uint32 MaxMappings)
 {
 #define ADDKEYMAP(KeyCode, KeyName)		if (NumMappings<MaxMappings) { KeyCodes[NumMappings]=KeyCode; KeyNames[NumMappings]=KeyName; ++NumMappings; };
